fix(test): correct the way a test is written
Changes the destination of the code I was writting in the test to the
Oferta constructor. #37
